docs: add docs and flag
